PR suggestions applied
