+++ engine/security/src/main/scala/pl/touk/nussknacker/ui/security/oauth2/OAuth2Configuration.scala
@@ -14,7 +14,7 @@ case class OAuth2Configuration(method: AuthenticationMethod,
                                clientSecret: String,
                                clientId: String,
                                profileUri: URI,
-                               profileFormat: ProfileFormat,
+                               profileFormat: Option[ProfileFormat],
                                accessTokenUri: URI,
                                redirectUri: URI,
                                implicitGrantEnabled: Boolean,
